cloud_functions/download_file_from_drive/requirements.txt
google-api-python-client==2.92.0
google-auth==2.23.2
google-cloud-storage==2.11.0
Flask==2.3.3
functions-framework==3.*
google-cloud-secret-manager==2.16.2
google-cloud-logging==3.11.2
requests==2.31.0